+++ tests/test_array.py
@@ -4,6 +4,7 @@ import time
 
 from .conftest import HqEnv
 from .utils import JOB_TABLE_ROWS, wait_for_job_state
+from .utils.job import default_task_output
 
 
 def test_job_array_submit(hq_env: HqEnv):
@@ -15,13 +16,27 @@ def test_job_array_submit(hq_env: HqEnv):
     wait_for_job_state(hq_env, 1, "FINISHED")
 
     for i in list(range(0, 30)) + list(range(37, 40)):
-        assert not os.path.isfile(os.path.join(hq_env.work_path, f"job-1/stdout.{i}"))
-        assert not os.path.isfile(os.path.join(hq_env.work_path, f"job-1/stderr.{i}"))
+        assert not os.path.isfile(
+            os.path.join(hq_env.work_path, default_task_output(job_id=1, task_id=i))
+        )
+        assert not os.path.isfile(
+            os.path.join(
+                hq_env.work_path,
+                default_task_output(job_id=1, task_id=i, type="stderr"),
+            )
+        )
 
     for i in range(36, 37):
-        stdout = os.path.join(hq_env.work_path, f"job-1/stdout.{i}")
+        stdout = os.path.join(
+            hq_env.work_path, default_task_output(job_id=1, task_id=i)
+        )
         assert os.path.isfile(stdout)
-        assert os.path.isfile(os.path.join(hq_env.work_path, f"job-1/stderr.{i}"))
+        assert os.path.isfile(
+            os.path.join(
+                hq_env.work_path,
+                default_task_output(job_id=1, task_id=i, type="stderr"),
+            )
+        )
         with open(stdout) as f:
             assert f.read() == f"1-{i}\n"